Improve readability of 'Makefile'

+++ make.sh
@@ -5,11 +5,17 @@ set -o errexit
 
 cd "$(dirname "${BASH_SOURCE[0]}")"
 
+# One can export variables PPL_SKIP_* before running this script to skip some parts.
+# Useful to spare some time on repeated runs.
+
 # Fail fast if Docker and NVidia's container runtime are not configured
-if ! docker run --rm --gpus all nvidia/cuda:11.2.2-base-ubuntu20.04 nvidia-smi >/dev/null 2>&1
+if [[ -z $PPL_SKIP_CHECK_GPU ]]
 then
-  echo "ERROR: Docker cannot use a CUDA-capable GPU. Please check the configuration of NVidia's container runtime"
-  exit 1
+  if ! docker run --rm --gpus all nvidia/cuda:11.2.2-base-ubuntu20.04 nvidia-smi >/dev/null 2>&1
+  then
+    echo "ERROR: Docker cannot use a CUDA-capable GPU. Please check the configuration of NVidia's container runtime"
+    exit 1
+  fi
 fi
 
 if [[ -z $PPL_SKIP_BUILDER ]]
@@ -19,7 +25,7 @@ fi
 
 docker run \
   --rm --interactive --tty \
-  --user $(id -u):$(id -g) \
+  --user $(id -u):$(id -g) `# Avoid creating files as 'root'` \
   --gpus all \
   --network none `# Ensure the repository is self-contained (except for the "docker build" phase)` \
   --volume "$PWD:/wd" --workdir /wd \
